Make every feature real end-to-end: docking, V2 UI surfaces, and verified data pipelines
- Docking works with real AutoDock Vina: strict PDB writer (no HEADER
  records), two-run pose comparison via ResultComparator over IPC
  (compare_results, list_jobs)
- All V2 surfaces wired in the UI: batch, water network, 2D editor,
  scripting console, comparison; real in-app modal replaces prompt()
- Contacts carry real atom names resolved from PLIP serials and group
  coordinates; measure_distance/angle resolve ligand atoms too
- run_script captures stdout so user scripts cannot corrupt IPC
- Batch reports failures honestly (no silent task drops)
- Real-data test suites: parser, contacts (live PLIP), enrichment (live
  APIs), docking (real Vina), v2; UI verification harness with screenshots
- snapcraft.yaml stages autodock-vina/plip/openbabel from distro packages
  plus the full GUI/runtime stack

+++ backend/ligora_backend/v2/comparison.py
@@ -2,8 +2,7 @@
 Result comparison for docking and analysis results.
 """
 
-from typing import Optional, List, Dict, Any, Tuple
-from dataclasses import dataclass, field
+from typing import Optional, List, Dict, Any
 from pathlib import Path
 
 import numpy as np
@@ -68,10 +67,10 @@ class ResultComparator:
                 comparisons.append({
                     'pose1': i + 1,
                     'pose2': j + 1,
-                    'rmsd': round(rmsd, 2),
+                    'rmsd': round(float(rmsd), 2),
                     'pose1_affinity': pose1.affinity,
                     'pose2_affinity': pose2.affinity,
-                    'energy_diff': round(energy_diff, 2),
+                    'energy_diff': round(float(energy_diff), 2),
                 })
 
         # Find best matches
@@ -83,7 +82,7 @@ class ResultComparator:
             'comparisons': comparisons,
             'best_matches': matches,
             'avg_rmsd': round(
-                np.mean([c['rmsd'] for c in comparisons]), 2
+                float(np.mean([c['rmsd'] for c in comparisons])), 2
             ) if comparisons else 0,
         }
 
@@ -109,7 +108,7 @@ class ResultComparator:
             sum_sq += dx*dx + dy*dy + dz*dz
 
         n = len(atoms1)
-        return np.sqrt(sum_sq / n) if n > 0 else 0.0
+        return float(np.sqrt(sum_sq / n)) if n > 0 else 0.0
 
     def _compute_superposed_rmsd(
         self,
@@ -172,7 +171,7 @@ class ResultComparator:
             diff = coords1_rotated - coords2_centered
             rmsd = np.sqrt(np.sum(diff * diff) / len(coords1))
 
-            return rmsd
+            return float(rmsd)
 
         except np.linalg.LinAlgError:
             return float('inf')
@@ -285,7 +284,7 @@ class ResultComparator:
         # Simple average
         avg_pose = DockingPose(
             pose_id=0,
-            affinity=np.mean([p.affinity for p in all_poses]),
+            affinity=float(np.mean([p.affinity for p in all_poses])),
             ligand_name=all_poses[0].ligand_name,
             atoms=[],
         )
